+++ src/types/fish.ts
@@ -2,6 +2,16 @@ import { EnergyHealth, ItemQuality, Season } from './common';
 
 export type FishCatchType = 'rod' | 'crab-pot';
 export type FishWeather = 'sunny' | 'rainy' | 'both';
+export type FishRoe = 'roe' | 'caviar';
+
+export interface FishPondProduce {
+  product: string;
+  minPopulation: number;
+}
+
+export interface FishPond {
+  produce: FishPondProduce[];
+}
 
 export interface Fish {
   id: string;
@@ -16,6 +26,8 @@ export interface Fish {
   sellPrice: number;
   fishTank: boolean;
   canSmoke: boolean;
+  roe: FishRoe | null;
+  fishPond: FishPond | null;
   usedIn: string[];
   energyHealth?: EnergyHealth;
   maxQuality: ItemQuality;